google_sheets_mcp/pyproject.toml
[project]
name = "google_sheets_mcp"
version = "0.1.0"
description = "MCP server for Google Sheets integration"
requires-python = ">=3.11"
dependencies = [
    "mcp>=0.1.0",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"